Make deck's `access` attribute editable and set default to 'private'
So far the `access` attribute was mostly a stub and not yet editable by
the user. Add a select element to the deck form to make it editable.
Also change the default `access` to private so that user collections are
private by default. Note that we don't plan to have private questions
currently, i.e. only the collection of questions is private but not the
contents of the individual questions in the deck.

Admins continue to be able to edit all decks, but we should add a
warning to decks where the admin user isn't the owner to avoid
accidental edits. (Will follow in a later commit.)

Also, add a few comments and extra lines to make the code more readable.

+++ app/Http/Controllers/Api/DeckQuestionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
 
 use App\Models\Deck;
@@ -12,12 +13,20 @@ class DeckQuestionController extends Controller
 {
     public function index(Deck $deck)
     {
+        abort_if($deck->access == "private" && $deck->user_id != Auth::id() && !Auth::user()->is_admin, 404);
+
         $questions = $deck->questions()->with('answers', 'images')->get();
+
         return response()->json($questions);
     }
 
     public function store(Request $request, Deck $deck)
     {
+        // Decks with access "private" or "public-ro" can only
+        // be updated by their owner or an admin
+        abort_if($deck->access == "private" && $deck->user_id != Auth::id() && !Auth::user()->is_admin, 404);
+        abort_if($deck->access == "public-ro" && $deck->user_id != Auth::id() && !Auth::user()->is_admin, 403);
+
         if ($request->id) {
             $question = Question::findOrFail($request->id)->get();
         } else {
@@ -39,6 +48,11 @@ class DeckQuestionController extends Controller
 
     public function destroy(Deck $deck, Question $question)
     {
+        // Decks with access "private" or "public-ro" can only
+        // be updated by their owner or an admin
+        abort_if($deck->access == "private" && $deck->user_id != Auth::id() && !Auth::user()->is_admin, 404);
+        abort_if($deck->access == "public-ro" && $deck->user_id != Auth::id() && !Auth::user()->is_admin, 403);
+
         $deck->questions()->detach($question->id);
 
         if ($question->decks->isEmpty()) {
